Use common helper class

+++ pj_tflite_pose_movenet/image_processor/pose_engine.h
@@ -40,7 +40,7 @@ public:
     typedef struct Result_ {
         std::vector<float>                                  pose_scores;			// [body]
         std::vector<std::vector<float>>                     pose_keypoint_scores;	// [body][part]
-        std::vector<std::vector<std::pair<float, float>>>   pose_keypoint_coords;	// [body][part][x, y] (0 - 1.0)
+        std::vector<std::vector<std::pair<int32_t, int32_t>>>   pose_keypoint_coords;	// [body][part][x, y]
         double    time_pre_process;		// [msec]
         double    time_inference;		// [msec]
         double    time_post_process;	// [msec]
